+++ src/store/blocks/Pid/actions.ts
@@ -1,5 +1,5 @@
 import { updateBlockState } from '../mutations';
-import { PidBlock } from './PID';
+import { PidBlock } from '@/store/blocks/Pid/Pid';
 
 import { RootStore } from '../../state';
